requirements.txt
django==2.2.6
pillow==7.0.0
pytz==2019.3
pytils==0.3
psycopg2-binary==2.8.6
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
requests==2.28.1
urllib3==1.26.11
python-dotenv==0.20.0
jsonrpclib==0.2.1
pycoin==0.92.20220529